pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-mistralai>=0.2.11",
    "langgraph>=0.5.2",
    "mistralai>=1.9.1",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
